pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "radiology-ai"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # --- Agent orchestration ---
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",

    # --- API ---
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-multipart>=0.0.9",

    # --- Medical imaging (torch/numpy/torchxrayvision installed separately in Dockerfile) ---
    "pydicom>=2.4.0",
    "Pillow>=10.3.0",

    # --- RAG / vector DB ---
    "qdrant-client>=1.9.0",
    "fastembed>=0.2.0",

    # --- Database ---
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",

    # --- AWS ---
    "boto3>=1.34.0",

    # --- MLOps ---
    "wandb>=0.17.0",

    # --- Observability ---
    "prometheus-client>=0.20.0",

    # --- HTTP client ---
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "packaging>=23.0",

    # --- Gradio UI ---
    "gradio>=4.36.0",

    # --- MCP ---
    "mcp>=1.0.0",

    # --- Utils ---
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "api*", "pipeline*", "ui*", "mlops*"]

[tool.ruff]
line-length = 100
